move build, train, scout orders files into agent scripted
