+++ nidup/pysc2/agent/scripted/scout.py
